actually, move the non-init stuff into core.js
